Fixing issue #5 - server should not connect to itself
This change is a bit more than just that though.  I've redesigned
(and hopefully simplified) the server by having it only take a single
argument, the host[:port] pairs of every server in the cluster.  Any
machine-local addresses are bound to and the rest considered peers.

This has the effect of only one server able to run per machine, even on
multiple network interfaces.  It's unpleasant for running a true
distributed test on a single machine, but in practice that never happens
and I think this will make things simpler.

The web context listener now checks the configuration and only fires up
a server if the machine running the web app is actually part of the
cluster.  Applications starting up on their own should do the same.

PubSubServer unit test now prints out timings, looks like on my laptop
there's roughly 35k/second message throughput for a message
[application] payload size of 88 bytes (this does not include the
overhead imposed by this system).  This includes the time to send a
message and have it re-broadcast to all other clients and handed back to
the application in de-serialized form.  Considering it's a laptop and it
was running inside Eclipse with Firefox and a screen saver running - not
too bad :)