Adding CapturingInputStream.
Wraps an input stream, while reading it writes to another stream as
well.  I created it because I was parsing a stream directly off of a
network connection yet wanted to save it to disk at the same time in a
local cache and didn't want to muck up my app code.  Seemed
generally useful.